+++ apps/consumer-client/components/Button.tsx
@@ -1,3 +1,0 @@
-export function Button() {
-  return <button>this is a button</button>;
-}