+++ apps/passport-client/src/model/PCDCollection.ts
@@ -1,9 +1,0 @@
-import { PCD } from "@pcd/pcd-types";
-
-export class PCDCollection {
-  private pcds: PCD[] = [];
-
-  public addPCD(pcd: PCD) {
-    this.pcds.push(pcd);
-  }
-}