+++ apps/passport-client/src/util.ts
@@ -1,0 +1,8 @@
+import { Dispatcher } from "./dispatch";
+
+export function err(dispatch: Dispatcher, title: string, message: string) {
+  dispatch({
+    type: "error",
+    error: { title, message },
+  });
+}