+++ packages/input-test-pcd/test/InputTestPCD.spec.ts
@@ -1,0 +1,3 @@
+describe("Semaphore Identity PCD", function () {
+  it("placeholder", async function () {});
+});